feat: Factory클래스로 이름 변경
